+++ src/pages/beta/pupil.tsx
@@ -3,17 +3,13 @@ import { FC } from "react";
 import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
 import Grid from "../../components/Grid";
 import GridArea from "../../components/Grid/GridArea";
-import Layout from "../../components/Layout";
+import AppLayout from "../../components/AppLayout";
 import MaxWidth from "../../components/MaxWidth/MaxWidth";
 import { Heading } from "../../components/Typography";
 
-const PupilHome: FC = () => {
+const PupilHomePage: FC = () => {
   return (
-    <Layout
-      seoProps={DEFAULT_SEO_PROPS}
-      headerVariant="app"
-      $background={"grey1"}
-    >
+    <AppLayout seoProps={DEFAULT_SEO_PROPS} $background={"grey1"}>
       <MaxWidth>
         <Grid $cg={16} $rg={[16, 48, 80]}>
           <GridArea $colSpan={[12, 12, 8]}>
@@ -28,8 +24,8 @@ const PupilHome: FC = () => {
           </GridArea>
         </Grid>
       </MaxWidth>
-    </Layout>
+    </AppLayout>
   );
 };
 
-export default PupilHome;
+export default PupilHomePage;